requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
openai==1.75.0
python-dotenv==1.0.1
pyinstaller==6.13.0